residual model api
